test(services): split round3 mutation sweeps into per-source files
Replace the six theme-grouped mutation_round3_*_test.go aggregates with
20 <source>_mutation_round3_test.go files matching the repo's
<source>_mutation[_roundN]_test.go convention, so each file names the one
production source it mutates. Shared stubs/helpers keep a single package-level
definition. Also drops the duplicate TestMR3I18n_LocalizedDateDisplayEnJanuaryYear999
(functionally identical to date_i18n_policy_mutation_test.go's January-999 test).
Behavior-only: 48 MR3 tests preserved, all green; no production code touched.
